LPS-74815 Move common parameters to Args class
